vysledky: Only refresh when online and make it off by default
When connecting the projector to internet over Wi-Fi hotspot,
the connection can drop when the phone gets out of range.

Let’s switch to refreshing using JavaScript and limit it to only when on-line.
And since it is now in JavaScript, let’s turn it off by default and add a menu item.